Unidas todas las funcionalidades

+++ Acme-BnB/src/main/java/controllers/lessor/LessorPropertyController.java
@@ -13,12 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-
-import services.LessorService;
 import services.PropertyService;
 import controllers.AbstractController;
-import domain.Lessor;
 import domain.Property;
+import forms.PropertyForm;
 
 @Controller
 @RequestMapping("/lessor/property")
@@ -29,8 +27,6 @@ public class LessorPropertyController extends AbstractController {
 	@Autowired
 	private PropertyService	propertyService;
 
-	@Autowired
-	private LessorService	lessorService;
 
 
 	//Constructor----------------------
@@ -49,9 +45,9 @@ public class LessorPropertyController extends AbstractController {
 
 		properties = propertyService.findByUserAccount();
 
-		result = new ModelAndView("tenant/list");
+		result = new ModelAndView("property/list");
 		result.addObject("properties", properties);
-		result.addObject("requestURI", "tenant/property/list.do");
+		result.addObject("requestURI", "lessor/property/list.do");
 
 		return result;
 	}
@@ -62,10 +58,10 @@ public class LessorPropertyController extends AbstractController {
 	public ModelAndView create() {
 
 		ModelAndView result;
-		Property property;
+		PropertyForm propertyForm;
 
-		property = propertyService.create();
-		result = createEditModelAndView(property);
+		propertyForm = propertyService.generateForm();
+		result = createEditModelAndView(propertyForm, null);
 
 		return result;
 
@@ -80,8 +76,10 @@ public class LessorPropertyController extends AbstractController {
 		Property property;
 
 		property = propertyService.findOne(propertyId);
+		//PropertyForm propertyForm = propertyService.transform(property);
 		Assert.notNull(property);
-		result = createEditModelAndView(property);
+		result = new ModelAndView("property/edit");
+		result.addObject("property", property);
 
 		return result;
 
@@ -90,17 +88,21 @@ public class LessorPropertyController extends AbstractController {
 	@RequestMapping(value = "/edit", method = RequestMethod.POST, params = "save")
 	public ModelAndView save(@Valid Property property, BindingResult binding) {
 
-		ModelAndView result;
-		Lessor lessor = lessorService.findByPrincipal();
-
-		if (binding.hasErrors() || property.getLessor() != lessor) {
+		ModelAndView result =  new ModelAndView();		
+		
+		if (binding.hasErrors()) {
 			result = createEditModelAndView(property);
 		} else {
 			try {
-				propertyService.save(property);
+				property = propertyService.reconstruct(property, binding);
+				propertyService.save2(property);
 				result = list();
 			} catch (Throwable oops) {
-				result = createEditModelAndView(property, "master.page.commit.error");
+				String msgCode = "lessor.register.error";
+				if (oops.getMessage().equals("nullRate")) {
+					msgCode = "lessor.property.nullRate";
+					result = createEditModelAndView(property, msgCode); 
+				}
 			}
 		}
 		return result;
@@ -110,34 +112,50 @@ public class LessorPropertyController extends AbstractController {
 	public ModelAndView delete(Property property, BindingResult binding) {
 
 		ModelAndView result;
-
-		try {
-			propertyService.delete(property);
-			result = list();
-		} catch (Throwable oops) {
-			result = createEditModelAndView(property, "master.page.commit.error");
+		
+		property = propertyService.reconstruct(property, binding);
+		if (binding.hasErrors()) {
+			result = createEditModelAndView(property);
+		} else {
+			try {
+				propertyService.delete(property);
+				result = list();
+			} catch (Throwable oops) {
+				result = createEditModelAndView(property);
+			}
 		}
 		return result;
 	}
 
 	//Ancillary Methods---------------------------
 
-	protected ModelAndView createEditModelAndView(Property property) {
-
-		ModelAndView result;
-
-		result = createEditModelAndView(property, null);
-
-		return result;
-	}
 
 	protected ModelAndView createEditModelAndView(Property property, String message) {
 		ModelAndView result;
 
 		result = new ModelAndView("property/edit");
 		result.addObject("property", property);
+		result.addObject("message", message);
+		return result;
+
+	}
+	
+	protected ModelAndView createEditModelAndView(PropertyForm property, String message) {
+		ModelAndView result;
+
+		result = new ModelAndView("property/edit");
+		result.addObject("property", property);
 
 		result.addObject("message", message);
+
+		return result;
+
+	}
+	
+	protected ModelAndView createEditModelAndView(Property property) {
+		ModelAndView result;
+
+		result = createEditModelAndView(property, null);
 
 		return result;
